Util functions to extract from and patch 2d arrays
The core of wet and sticky relies on analysing cells along side their
neighbours so I needed to create a function to extract a smaller 2d
array from inside a larger array.

Once the smaller array has been operated on it would need to be patched
back into the larger array, returning the new state of the array. I have
implemented a patch function to do this.