Add POC engine to GUI and cycle to core
As part of wet and sticky we need to have an engine that constantly runs
an algorithm on random cells. This adds a thread that will repeatedly
pick a random i j cell and run a cycle on it. At the moment the cycle
just changes the colour of the cell to red so that you can see it
working.