Create ClojureScript output for paint
The HSLMixer tests had to change due to ClojureScripts lack of ratio
support. I've switched them to converting everything to floats and then
comparing them with some error.

Macros can only be clj so we have a new namespace where these will live.

The current API of paint has been exported so that when building with
advanced optimizations everything doesn't get shaken away and so that
external users can access the functions.

This also includes a page to host the outputted javascript. It will be
expanded to demonstrate the library in action.