Create substrate with w, h and attributes
Previously if you wanted to provide custom attributes for the substrate
then you would also have to provide a mixer and chronos. This was a bit
of a pita so I've added another signature that will accept just a width,
height and attribute map still defaulting the mixer and chronos.